backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 
opentelemetry-instrumentation==0.55b0
opentelemetry-distro==0.55b0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-instrumentation-asyncio==0.55b0
opentelemetry-instrumentation-dbapi==0.55b0
opentelemetry-instrumentation-logging==0.55b0
opentelemetry-instrumentation-sqlite3==0.55b0
opentelemetry-instrumentation-threading==0.55b0
opentelemetry-instrumentation-urllib==0.55b0
opentelemetry-instrumentation-wsgi==0.55b0
opentelemetry-instrumentation-click==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-instrumentation-grpc==0.55b0
opentelemetry-instrumentation-httpx==0.55b0
opentelemetry-instrumentation-requests==0.55b0
opentelemetry-instrumentation-starlette==0.55b0
opentelemetry-instrumentation-tortoiseorm==0.55b0
opentelemetry-instrumentation-urllib3==0.55b0
email-validator==2.1.0
pydantic-settings==2.1.0
